Added FACT as ProjectType and renamed the 'standard' one 'ARTEMIS_C'

+++ src/main/java/de/tum/in/www1/artemis/service/connectors/bamboo/BambooBuildPlanService.java
@@ -198,16 +198,21 @@ public class BambooBuildPlanService {
                 return createDefaultStage(programmingLanguage, sequentialBuildRuns, checkoutTask, defaultStage, defaultJob, "test-reports/*results.xml");
             }
             case C -> {
-                // Default tasks:
-                var tasks = readScriptTasksFromTemplate(programmingLanguage, "", sequentialBuildRuns, false, null);
-                tasks.add(0, checkoutTask);
-                defaultJob.tasks(tasks.toArray(new Task[0]));
-                // Final tasks:
-                final TestParserTask testParserTask = new TestParserTask(TestParserTaskProperties.TestType.JUNIT).resultDirectories("test-reports/*results.xml");
-                final ScriptTask cleanupTask = new ScriptTask().description("cleanup").inlineBody("sudo rm -rf tests/\nsudo rm -rf assignment/\nsudo rm -rf test-reports/");
-                defaultJob.finalTasks(testParserTask, cleanupTask);
-                defaultStage.jobs(defaultJob);
-                return defaultStage;
+                if (ProjectType.FACT.equals(projectType)) {
+                    throw new UnsupportedOperationException("FACT template is not available for Bamboo yet.");
+                }
+                else {
+                    // Default tasks:
+                    var tasks = readScriptTasksFromTemplate(programmingLanguage, "", sequentialBuildRuns, false, null);
+                    tasks.add(0, checkoutTask);
+                    defaultJob.tasks(tasks.toArray(new Task[0]));
+                    // Final tasks:
+                    final TestParserTask testParserTask = new TestParserTask(TestParserTaskProperties.TestType.JUNIT).resultDirectories("test-reports/*results.xml");
+                    final ScriptTask cleanupTask = new ScriptTask().description("cleanup").inlineBody("sudo rm -rf tests/\nsudo rm -rf assignment/\nsudo rm -rf test-reports/");
+                    defaultJob.finalTasks(testParserTask, cleanupTask);
+                    defaultStage.jobs(defaultJob);
+                    return defaultStage;
+                }
             }
             case HASKELL, OCAML -> {
                 return createDefaultStage(programmingLanguage, sequentialBuildRuns, checkoutTask, defaultStage, defaultJob, "**/test-reports/*.xml");
@@ -374,7 +379,7 @@ public class BambooBuildPlanService {
     }
 
     private DockerConfiguration dockerConfigurationImageNameFor(ProgrammingLanguage language) {
-        var dockerImage = getDockerImageName(language);
+        var dockerImage = getDockerImageName(language, null);
         return new DockerConfiguration().image(dockerImage);
     }
 }
